+++ lib/money/money.rb
@@ -1,7 +1,5 @@
 # encoding: utf-8
-require 'money/bank/variable_exchange'
 require 'money/money/arithmetic'
-require 'money/money/parsing'
 require 'money/money/formatting'
 
 # Represents an amount of money in a given currency.
@@ -9,12 +7,13 @@ class Money
   include Comparable
   include Arithmetic
   include Formatting
-  include Parsing
 
-  # The value of the money in cents.
+  # The value of the money.
   #
-  # @return [Integer]
-  attr_reader :cents
+  # @return [BigDecimal
+  def amount
+    @amount
+  end
 
   # The currency the money is in.
   #
@@ -54,8 +53,6 @@ class Money
     attr_accessor :assume_from_symbol
   end
 
-  # Set the default bank for creating new +Money+ objects.
-  self.default_bank = Bank::VariableExchange.instance
 
   # Set the default currency for creating new +Money+ object.
   self.default_currency = Currency.new("USD")
@@ -122,51 +119,6 @@ class Money
     Money.new(cents, "EUR")
   end
 
-  # Creates a new Money object of +amount+ value in dollars,
-  # with given +currency+.
-  #
-  # The amount value is expressed in +dollars+
-  # where the +dollar+ is the main monetary unit,
-  # opposite to the subunit-based representation
-  # used internally by this library called +cents+.
-  #
-  # @param [Numeric] amount The money amount, in dollars.
-  # @param [Currency, String, Symbol] currency The currency format.
-  # @param [Money::Bank::*] bank The exchange bank to use.
-  #
-  # @return [Money]
-  #
-  # @example
-  #   Money.new_with_dollars(100)
-  #   #=> #<Money @cents=10000 @currency="USD">
-  #   Money.new_with_dollars(100, "USD")
-  #   #=> #<Money @cents=10000 @currency="USD">
-  #   Money.new_with_dollars(100, "EUR")
-  #   #=> #<Money @cents=10000 @currency="EUR">
-  #
-  # @see Money.new
-  #
-  def self.new_with_dollars(amount, currency = Money.default_currency, bank = Money.default_bank)
-    money = from_numeric(amount, currency)
-    # Hack! You can't change a bank
-    money.instance_variable_set("@bank", bank)
-    money
-  end
-
-  # Adds a new exchange rate to the default bank and return the rate.
-  #
-  # @param [Currency, String, Symbol] from_currency Currency to exchange from.
-  # @param [Currency, String, Symbol] to_currency Currency to exchange to.
-  # @param [Numeric] rate Rate to exchange with.
-  #
-  # @return [Numeric]
-  #
-  # @example
-  #   Money.add_rate("USD", "CAD", 1.25) #=> 1.25
-  def self.add_rate(from_currency, to_currency, rate)
-    Money.default_bank.add_rate(from_currency, to_currency, rate)
-  end
-
 
   # Creates a new Money object of +cents+ value in cents,
   # with given +currency+.
@@ -190,10 +142,9 @@ class Money
   #
   # @see Money.new_with_dollars
   #
-  def initialize(cents, currency = Money.default_currency, bank = Money.default_bank)
-    @cents = cents.round.to_i
+  def initialize(amount, currency = Money.default_currency)
+    @amount = BigDecimal(amount.to_s)
     @currency = Currency.wrap(currency)
-    @bank = bank
   end
 
   # Returns the value of the money in dollars,
@@ -242,7 +193,7 @@ class Money
   # @example
   #   Money.new(100).hash #=> 908351
   def hash
-    [cents.hash, currency.hash].hash
+    [amount.hash, currency.hash].hash
   end
 
   # Uses +Currency#symbol+. If +nil+ is returned, defaults to "¤".
@@ -259,7 +210,7 @@ class Money
   #
   # @return [String]
   def inspect
-    "#<Money cents:#{cents} currency:#{currency}>"
+    "#<Money amount:#{amount} currency:#{currency}>"
   end
 
   # Returns the amount of money as a string.
@@ -269,14 +220,7 @@ class Money
   # @example
   #   Money.ca_dollar(100).to_s #=> "1.00"
   def to_s
-    unit, subunit  = cents.abs.divmod(currency.subunit_to_unit).map{|o| o.to_s}
-    if currency.decimal_places == 0
-      return "-#{unit}" if cents < 0
-      return unit
-    end
-    subunit = (("0" * currency.decimal_places) + subunit)[(-1*currency.decimal_places)..-1]
-    return "-#{unit}#{decimal_mark}#{subunit}" if cents < 0
-    "#{unit}#{decimal_mark}#{subunit}"
+    amount.to_s
   end
 
   # Return the amount of money as a BigDecimal.
@@ -286,7 +230,7 @@ class Money
   # @example
   #   Money.us_dollar(100).to_d => BigDecimal.new("1.0")
   def to_d
-    BigDecimal.new(cents.to_s) / BigDecimal.new(currency.subunit_to_unit.to_s)
+    amount
   end
 
   # Return the amount of money as a float. Floating points cannot guarantee
@@ -299,121 +243,7 @@ class Money
   # @example
   #   Money.us_dollar(100).to_f => 1.0
   def to_f
-    to_d.to_f
+    amount.to_f
   end
-
-  # Conversation to +self+.
-  #
-  # @return [self]
-  def to_money(given_currency = nil)
-    given_currency = Currency.wrap(given_currency) if given_currency
-    if given_currency.nil? || self.currency == given_currency
-      self
-    else
-      exchange_to(given_currency)
-    end
-  end
-
-  # Receive the amount of this money object in another Currency.
-  #
-  # @param [Currency, String, Symbol] other_currency Currency to exchange to.
-  #
-  # @return [Money]
-  #
-  # @example
-  #   Money.new(2000, "USD").exchange_to("EUR")
-  #   Money.new(2000, "USD").exchange_to(Currency.new("EUR"))
-  def exchange_to(other_currency)
-    other_currency = Currency.wrap(other_currency)
-    @bank.exchange_with(self, other_currency)
-  end
-
-  # Receive a money object with the same amount as the current Money object
-  # in american dollars.
-  #
-  # @return [Money]
-  #
-  # @example
-  #   n = Money.new(100, "CAD").as_us_dollar
-  #   n.currency #=> #<Money::Currency id: usd>
-  def as_us_dollar
-    exchange_to("USD")
-  end
-
-  # Receive a money object with the same amount as the current Money object
-  # in canadian dollar.
-  #
-  # @return [Money]
-  #
-  # @example
-  #   n = Money.new(100, "USD").as_ca_dollar
-  #   n.currency #=> #<Money::Currency id: cad>
-  def as_ca_dollar
-    exchange_to("CAD")
-  end
-
-  # Receive a money object with the same amount as the current Money object
-  # in euro.
-  #
-  # @return [Money]
-  #
-  # @example
-  #   n = Money.new(100, "USD").as_euro
-  #   n.currency #=> #<Money::Currency id: eur>
-  def as_euro
-    exchange_to("EUR")
-  end
-
-  # Allocates money between different parties without loosing pennies.
-  # After the mathmatically split has been performed, left over pennies will
-  # be distributed round-robin amongst the parties. This means that parties
-  # listed first will likely recieve more pennies then ones that are listed later
-  #
-  # @param [0.50, 0.25, 0.25] to give 50% of the cash to party1, 25% ot party2, and 25% to party3.
-  #
-  # @return [Array<Money, Money, Money>]
-  #
-  # @example
-  #   Money.new(5, "USD").allocate([0.3,0.7)) #=> [Money.new(2), Money.new(3)]
-  #   Money.new(100, "USD").allocate([0.33,0.33,0.33]) #=> [Money.new(34), Money.new(33), Money.new(33)]
-  def allocate(splits)
-    allocations = splits.inject(0.0) {|sum, i| sum += i }
-    raise ArgumentError, "splits add to more then 100%" if (allocations - 1.0) > Float::EPSILON
-
-    left_over = cents
-
-    amounts = splits.collect do |ratio|
-      fraction = (cents * ratio / allocations).floor
-      left_over -= fraction
-      fraction
-    end
-
-    left_over.times { |i| amounts[i % amounts.length] += 1 }
-
-    amounts.collect { |cents| Money.new(cents, currency) }
-  end
-
-  # Split money amongst parties evenly without loosing pennies.
-  #
-  # @param [2] number of parties.
-  #
-  # @return [Array<Money, Money, Money>]
-  #
-  # @example
-  #   Money.new(100, "USD").split(3) #=> [Money.new(34), Money.new(33), Money.new(33)]
-  def split(num)
-    raise ArgumentError, "need at least one party" if num < 1
-    low = Money.new(cents / num)
-    high = Money.new(low.cents + 1)
-
-    remainder = cents % num
-    result = []
-
-    num.times do |index|
-      result[index] = index < remainder ? high : low
-    end
-
-    result
-  end
-
+  
 end