+++ lib/money/money/arithmetic.rb
@@ -6,9 +6,9 @@ class Money
     # @return [Money]
     #
     # @example
-    #    - Money.new(100) #=> #<Money @cents=-100>
+    #    - Money.new(100) #=> #<Money @amount=-100>
     def -@
-      Money.new(-cents, currency)
+      Money.new(-amount, currency)
     end
 
 
@@ -26,7 +26,7 @@ class Money
     def ==(other_money)
       if other_money.respond_to?(:to_money)
         other_money = other_money.to_money
-        cents == other_money.cents && self.currency == other_money.currency
+        amount == other_money.amount && self.currency == other_money.currency
       else
         false
       end
@@ -47,9 +47,9 @@ class Money
       if other_money.respond_to?(:to_money)
         other_money = other_money.to_money
         if self.currency == other_money.currency
-          cents <=> other_money.cents
+          amount <=> other_money.amount
         else
-          cents <=> other_money.exchange_to(currency).cents
+          amount <=> other_money.exchange_to(currency).amount
         end
       else
         raise ArgumentError, "Comparison of #{self.class} with #{other_money.inspect} failed"
@@ -66,7 +66,7 @@ class Money
     #   Money.new(0).positive?  #=> false
     #   Money.new(-1).positive? #=> false
     def positive?
-      cents > 0
+      amount > 0
     end
 
     # Test if the amount is negative. Returns +true+ if the money amount is
@@ -79,7 +79,7 @@ class Money
     #   Money.new(0).negative?  #=> false
     #   Money.new(1).negative?  #=> false
     def negative?
-      cents < 0
+      amount < 0
     end
 
     # Returns a new Money object containing the sum of the two operands' monetary
@@ -91,12 +91,12 @@ class Money
     # @return [Money]
     #
     # @example
-    #   Money.new(100) + Money.new(100) #=> #<Money @cents=200>
+    #   Money.new(100) + Money.new(100) #=> #<Money @amount=200>
     def +(other_money)
       if currency == other_money.currency
-        Money.new(cents + other_money.cents, other_money.currency)
+        Money.new(amount + other_money.amount, other_money.currency)
       else
-        Money.new(cents + other_money.exchange_to(currency).cents, currency)
+        Money.new(amount + other_money.exchange_to(currency).amount, currency)
       end
     end
 
@@ -110,12 +110,12 @@ class Money
     # @return [Money]
     #
     # @example
-    #   Money.new(100) - Money.new(99) #=> #<Money @cents=1>
+    #   Money.new(100) - Money.new(99) #=> #<Money @amount=1>
     def -(other_money)
       if currency == other_money.currency
-        Money.new(cents - other_money.cents, other_money.currency)
+        Money.new(amount - other_money.amount, other_money.currency)
       else
-        Money.new(cents - other_money.exchange_to(currency).cents, currency)
+        Money.new(amount - other_money.exchange_to(currency).amount, currency)
       end
     end
 
@@ -131,13 +131,13 @@ class Money
     # @raise [ArgumentError] If +value+ is a Money instance.
     #
     # @example
-    #   Money.new(100) * 2 #=> #<Money @cents=200>
+    #   Money.new(100) * 2 #=> #<Money @amount=200>
     #
     def *(value)
       if value.is_a?(Money)
         raise ArgumentError, "Can't multiply a Money by a Money"
       else
-        Money.new(cents * value, currency)
+        Money.new(amount * value, currency)
       end
     end
 
@@ -153,18 +153,18 @@ class Money
     # @return [Float] The resulting number if you divide Money by a Money.
     #
     # @example
-    #   Money.new(100) / 10            #=> #<Money @cents=10>
+    #   Money.new(100) / 10            #=> #<Money @amount=10>
     #   Money.new(100) / Money.new(10) #=> 10.0
     #
     def /(value)
       if value.is_a?(Money)
         if currency == value.currency
-          (cents / BigDecimal.new(value.cents.to_s)).to_f
+          (amount / BigDecimal.new(value.amount.to_s)).to_f
         else
-          (cents / BigDecimal(value.exchange_to(currency).cents.to_s)).to_f
+          (amount / BigDecimal(value.exchange_to(currency).amount.to_s)).to_f
         end
       else
-        Money.new(cents / value, currency)
+        Money.new(amount / value, currency)
       end
     end
 
@@ -189,16 +189,16 @@ class Money
     # @return [Array<Money,Money>,Array<Fixnum,Money>]
     #
     # @example
-    #   Money.new(100).divmod(9)            #=> [#<Money @cents=11>, #<Money @cents=1>]
-    #   Money.new(100).divmod(Money.new(9)) #=> [11, #<Money @cents=1>]
+    #   Money.new(100).divmod(9)            #=> [#<Money @amount=11>, #<Money @amount=1>]
+    #   Money.new(100).divmod(Money.new(9)) #=> [11, #<Money @amount=1>]
     def divmod(val)
       if val.is_a?(Money)
-        a = self.cents
-        b = self.currency == val.currency ? val.cents : val.exchange_to(self.currency).cents
+        a = self.amount
+        b = self.currency == val.currency ? val.amount : val.exchange_to(self.currency).amount
         q, m = a.divmod(b)
         return [q, Money.new(m, self.currency)]
       else
-        return [self.div(val), Money.new(self.cents.modulo(val), self.currency)]
+        return [self.div(val), Money.new(self.amount.modulo(val), self.currency)]
       end
     end
 
@@ -209,8 +209,8 @@ class Money
     # @return [Money]
     #
     # @example
-    #   Money.new(100).modulo(9)            #=> #<Money @cents=1>
-    #   Money.new(100).modulo(Money.new(9)) #=> #<Money @cents=1>
+    #   Money.new(100).modulo(9)            #=> #<Money @amount=1>
+    #   Money.new(100).modulo(Money.new(9)) #=> #<Money @amount=1>
     def modulo(val)
       self.divmod(val)[1]
     end
@@ -233,14 +233,14 @@ class Money
     # @return [Money]
     #
     # @example
-    #   Money.new(100).remainder(9) #=> #<Money @cents=1>
+    #   Money.new(100).remainder(9) #=> #<Money @amount=1>
     def remainder(val)
       a, b = self, val
       b = b.exchange_to(a.currency) if b.is_a?(Money) and a.currency != b.currency
 
       a_sign, b_sign = :pos, :pos
-      a_sign = :neg if a.cents < 0
-      b_sign = :neg if (b.is_a?(Money) and b.cents < 0) or (b < 0)
+      a_sign = :neg if a.amount < 0
+      b_sign = :neg if (b.is_a?(Money) and b.amount < 0) or (b < 0)
 
       return a.modulo(b) if a_sign == b_sign
       a.modulo(b) - (b.is_a?(Money) ? b : Money.new(b, a.currency))
@@ -251,9 +251,9 @@ class Money
     # @return [Money]
     #
     # @example
-    #   Money.new(-100).abs #=> #<Money @cents=100>
+    #   Money.new(-100).abs #=> #<Money @amount=100>
     def abs
-      Money.new(self.cents.abs, self.currency)
+      Money.new(self.amount.abs, self.currency)
     end
 
     # Test if the money amount is zero.
@@ -264,7 +264,7 @@ class Money
     #   Money.new(100).zero? #=> false
     #   Money.new(0).zero?   #=> true
     def zero?
-      cents == 0
+      amount == 0
     end
 
     # Test if the money amount is non-zero. Returns this money object if it is
@@ -273,10 +273,10 @@ class Money
     # @return [Money, nil]
     #
     # @example
-    #   Money.new(100).nonzero? #=> #<Money @cents=100>
+    #   Money.new(100).nonzero? #=> #<Money @amount=100>
     #   Money.new(0).nonzero?   #=> nil
     def nonzero?
-      cents != 0 ? self : nil
+      amount != 0 ? self : nil
     end
 
   end